add adminjs and typeorm config
